Miglioramenti su registrazione incosso rata
1. Miglioramwento struttura database e risolte incongruenze
2. Risolti alcuni errori di migration
3. Migliorata logica creazione nuovo condomino con la creazione dei mastri dei conti contabili e aggiunta middleware per compatibilità con versioni già installate
4. Migliorata gerarchia database quando si crea una nuiova cassa

+++ database/migrations/2025_12_17_213207_add_conto_contabile_id_to_conti_table.php
@@ -8,20 +8,28 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('conti', function (Blueprint $table) {
-            $table->foreignId('conto_contabile_id')
-                ->nullable()
-                ->after('parent_id')
-                ->constrained('conti_contabili')
-                ->nullOnDelete();
-        });
+       // Esegui la modifica SOLO se la colonna non esiste ancora
+        if (Schema::hasTable('conti') && !Schema::hasColumn('conti', 'conto_contabile_id')) {
+            Schema::table('conti', function (Blueprint $table) {
+                $table->foreignId('conto_contabile_id')
+                    ->nullable()
+                    ->after('parent_id')
+                    ->constrained('conti_contabili')
+                    ->nullOnDelete();
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('conti', function (Blueprint $table) {
-            $table->dropForeign(['conto_contabile_id']);
-            $table->dropColumn('conto_contabile_id');
-        });
+        // Controlliamo se la tabella e la colonna esistono prima di provare a cancellarle
+        if (Schema::hasTable('conti') && Schema::hasColumn('conti', 'conto_contabile_id')) {
+            Schema::table('conti', function (Blueprint $table) {
+                // È fondamentale rimuovere PRIMA il vincolo (foreign key)
+                $table->dropForeign(['conto_contabile_id']);
+                // E POI la colonna
+                $table->dropColumn('conto_contabile_id');
+            });
+        }
     }
 };
